+++ docs/docs/.vitepress/utils/sidebar.ts
@@ -55,7 +55,8 @@ export const sidebar = {
         { text: 'Switch 开关', link: '/components/switch' },
         { text: 'Avatar 头像', link: '/components/avatar' },
         { text: 'Progress 进度条', link: '/components/progress' },
-        { text: 'Empty 空状态', link: '/components/empty' }
+        { text: 'Empty 空状态', link: '/components/empty' },
+        { text: 'List 列表', link: '/components/list' }
       ]
     },
     {