+++ packages/fighting-components/list/src/list.vue
@@ -1,7 +1,7 @@
 <script lang="ts" setup name="FList">
-  import type { ComputedRef, CSSProperties } from 'vue'
   import { provide, computed } from 'vue'
   import { Props, listPropsKey } from './list'
+  import type { ComputedRef, CSSProperties } from 'vue'
   import type { FPropsType } from './list'
 
   const props: FPropsType = defineProps(Props)
@@ -10,34 +10,35 @@
 
   const listClass: ComputedRef<object | string[]> = computed(
     (): object | string[] => {
-      const { maxHeight, split, showNum, showDisc } = props
+      const { maxHeight, zebra, center, size } = props
 
       return [
         'f-list',
+        `f-list-${size}`,
         {
           'f-list-scroll': maxHeight,
-          'f-list-split': split,
-          'f-list-show-number': !showDisc && showNum,
-          'f-list-show-disc': !showNum && showDisc
+          'f-list-zebra': zebra,
+          'f-list-center': center
         }
       ] as const
     }
   )
 
   const listStyle: ComputedRef<CSSProperties> = computed((): CSSProperties => {
-    const { maxHeight, borderColor } = props
+    const { maxHeight, borderColor, listStyle } = props
 
     return {
       maxHeight,
-      borderColor
+      borderColor,
+      listStyle
     } as const
   })
 </script>
 
 <template>
-  <ol :class="listClass" :style="listStyle">
+  <ul :class="listClass" :style="listStyle">
     <slot name="header" />
     <slot />
     <slot name="footer" />
-  </ol>
+  </ul>
 </template>