+++ start/src/App.vue
@@ -1,7 +1,5 @@
 <script lang="ts" setup></script>
 
-<template>
-  <f-avatar src="https://tianyuh2ao.cn/images/auto/my.jpg" />
-</template>
+<template></template>
 
 <style scoped></style>