+++ start/src/App.vue
@@ -1,5 +1,7 @@
 <script lang="ts" setup></script>
 
-<template></template>
+<template>
+  <f-avatar :size="85" src="https://tianyuhao.cn/images/auto/my.jpg" />
+</template>
 
 <style scoped></style>