+++ packages/fighting-components/_hooks/index.ts
@@ -1,1 +1,0 @@
-export { useFilterProps } from './useFilterProps'