+++ packages/fighting-components/_interface/index.d.ts
@@ -1,4 +1,7 @@
-export type { RipplesInterface, RipplesNeedButtonPropsInterface } from './button'
+export type {
+  RipplesInterface,
+  RipplesNeedButtonPropsInterface
+} from './button'
 export type {
   LazyInterface,
   loadImageInterface,
@@ -15,5 +18,8 @@ export type {
   pastTimeConfigInterface
 } from './utils'
 export type { changeColorInterface } from './change-color'
-export type { createBase64Interface, createBase64NeedWatermarkPropsInterface } from './watermark'
+export type {
+  createBase64Interface,
+  createBase64NeedWatermarkPropsInterface
+} from './watermark'
 export type { ordinaryFunctionInterface } from './auto'