+++ packages/fighting-components/button/src/button.vue
@@ -3,7 +3,7 @@
   import { computed, ref } from 'vue'
   import { Props, Emits } from './button'
   import { Ripples, ChangeColor } from '../../_utils'
-  import { useFilterProps } from '../../_hooks'
+  import { useFilterProps } from '../../_hooks/useFilterProps'
   import type { FPropsType } from './button'
   import type { ComputedRef, Ref, CSSProperties } from 'vue'
   import type { onClickInterface as a } from './interface'